requirements.txt
async-timeout==4.0.2
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2r
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.0.33
protobuf==3.20.3
redis==4.6.0
requests==2.31.0
urllib3==2.0.3
Werkzeug==2.3.6
